UNTESTED: added a new deep search mode for one-time linedefs
DeepCheckLosses. This should prevent the bot from activating lines if
they result in a loss of access to current possible destinations. This
should prevent the E1M3 death dive at the start.

This check should really be done for all kinds of destinations, not just
one-time linedefs, but it's something TODO.

Not tested.