Now calling Check_Sides in A_PainShootSkull is line-portal-aware
Whereas P_SpawnMobj is called for trans-portal coordinates, Check_Sides
will only work if the two coordinates are close together, as is the case
when summoning anything in a map without portals, so it will just use
the initial, untranslated offset, taking advantage of the fact that
portal lines (behind which these coordinates would result) are
necessarily passable.

Spawning on intersection with sector portals still doesn't work when the
"lost souls stuck behind walls" compatibility is set to "NO", as in MBF,
but this is a separate issue.