Portal_Define has been updated, and portalfloor and portalceiling UDMF sector fields exist
Now for linked portals, the Z parameter offsets the plane Z. The plane Z is decided based on the two lines' front sectors. It's at the trigger line's front sector floor if the target sector is lower or equal, or at trigger line's ceiling sector if the target sector is higher. By height here I mean average between floorheight and ceilingheight.

More importantly, for any two-way or linked portal, two portals are created instead of one! The second portal has a negative of the provided ID, and a reversed rendering offset, meant to be used on the opposite side.