Partially fixed unwanted walls and sprites appearing on top of anchored non-linked ceiling/floor portals
It uses a rough delimitation based on bounding box of the target area. Anything inside that bounding box may still be rendered even if not desired, but if you keep external architecture out of that box, it will be rejected.

Linked portals don't have this problem: segs and sprites belonging to different group IDs are already rejected.