Fixed the glitches with edge portals caused by bad seg culling
I did it by adding some slack offset to the tested seg coordinates so if they are on top of the portal line, they'll actually appear as "behind" (towards you), so they're rejected. This prevents infinite portal recursions when looking down/up through the edge starting from the sector portal.