Vertex slopes are now supported in UDMF maps
For each triangular sector having vertex-defined slopes, the sector height will also be changed to the average (centroid) height, so you can just set slopes in the editor, and you'll already have a pretty good approximation of the slope heights, even if accurate clipping is not supported yet.

The downside to this is that logical sector heights (in relation to stuff such as plat specials) will also change, becoming quite incompatible with ZDoom-like designs. And the already-working ZDoom slope physics is cleaner as it is (editor-defined floor height with added/removed slope effects). This is the reason I didn't commit this directly on master.