Started overlaid automap.
Currently it works by cycling with the tab key between modes (off, overlaid and full). The only currently known bug is that when removing the status bar, the automap is still cut by it. Ideally it should update its visible range and centre when user changes view, but that means reinitializing some variables.