Fixed the bug where portal overlays would get overdrawn
It was because they were associated with portals and not windows. Portals can appear in multiple windows, so they would get listed too many times in the post-BSP stack.

The trick to work with this is to keep overlay sets in portals as a temporary store (just so visplanes can be created for the seg clipping logic to work without heavy changes), but when the window is created, or referenced, the visplanes are also stored in its overlay set.

Since windows die too early, their overlays need to be stored independently. They're only cleared when the stack stop needing them, though they may also stay in the windows if unused.

Child windows don't spawn overlays.