New EDF thingtype property: projectilealliance
It allows monsters of different species but with projectiles to not hurt each other. It seeks to reproduce the hell knight and baron of hell alliance.

Known problem: now, every monster who inherits the baron or knight will be part of the same alliance (unless I null out the property). This is going to contradict past mods which used monsters based on hell knights.