When playing under vanilla mode, emulate the way it handles crushing stairs.
In Doom, it's undefined behaviour which happens to be most of the time in the way the designer expects: stairs almost always crush and pause movement. Sometimes they may crush but move, or not crush at all, but those cases are rare and not expected by designers. Blatant example which needs this: Cyberdreams. This is not a problem when simply playing Cyberdreams under modern compatibility settings, but is needed for playing vanilla demos.