sort+binary search for drawsegs obscuring sprites
optimization for bad-case of many drawsegs in R_DrawSpriteInDSRange &
R_DrawPostBSP. we sort the drawsegs, then use a binarysearch to set the
starting point in drawsegs_xrange to check per sprite. this saves a lot
of checks vs the original linear pass through the array, assuming there
are many drawsegs. in tests on maps like The Given, i had a 5-10fps
increase (up to 25-30 total). there is a drawback: additional overhead
from the sort and the binary search as well as some additional checks
later in the code (to compensate for messing up the previous order),
means that we lose some speed in very simple scenes. In testing, this
amounts to  losing 1-2 fps on scenes that are already 200+fps