CHANGED the way portal polyobjects move relative to objects
All airborne (and not hanging from ceiling) objects will remain relative to the ground. This may be less physically realistic, but is more consistent with how traditional up-down platforms work in Doom relative to things, and is also more robust, as it avoids ambiguous situations where floating projectiles would weirdly move with the platform. Also, the previous way of adapting velocity on passing portal didn't always work as expected.