Don't reset references to 0 in addThinker
This mustn't be done, because it's possible "references" was already raised by a call to P_SetTarget executed before the current thinker got installed in the list with addThinker. Snatching references like that is a sure way to memory leaks or crashes.

In practice this happened when an object had alphavelocity, thus producing a MobjFadeThinker which targets this Mobj, before its addThinker method got called.

It used to be necessary in Boom or MBF to reset references to 0 in addThinker because then there was no RAII, but now we safely set references to 0 in Thinker's constructor.