Hopefully fix a crash happening sometimes when loading maps with MobjFadeThinkers
It happens if you save during this transition: a Mobj gets removed. Then the game is saved. During saving, a MobjFadeThinker referring to that removed Mobj (which would get removed during gameplay ticking) now has a removed reference, which has to be saved as NULL. When loading that game, we get a MobjFadeThinker with a NULL Mobj reference, which is unexpected and will crash. This update makes sure that such a MobjFadeThinker gets removed the same way as with a non-NULL but removed reference.

However, I'm getting different crashes and I'm not sure whether they're linked.