Only use the first like-named entries from TEXTURE1/TEXTURE2
If there are multiple textures with the same name in TEXTURE1+TEXTURE2, only add the first one to the hash table. Needed because certain mods, such as Army of Darkness DOOM, have this problem, and they were looking wrong in Eternity.

Other textures are loaded correctly. For example if you have some IWAD texture replaced with a lump between TX_START and TX_END, the new lump will be used.