make PatchLoader::CacheNum safe by returning default patch instead of nullptr
this simplifies CacheName down to one line. additionally, none of the uses of
this function assume that CacheNum might return nullptr, so this is a safe
change to make. all current uses of CacheNum either will crash upon invalid
use, or already check if the lump exists before calling CacheNum. this fixes
those crashes and does not affect anything else.