Hopefully fixed the occasional failure of extending a sector portal with front continuation segs
This happened if a front seg was merely a continuation of a back sector-windowed seg. It would just momentarily get a sector window pointer just for some other (but similar) maintenance, then it would be nulled and ultimately not found by the function responsible for extending the render barrier bounding box. I've moved this function closer to the sector window assignment, to make sure it's called.