Managed to preserve Forseti demo compatibility by adding 2 lines of code
Sorry, but it was easy enough to add. Still, maybe we still shouldn't consider it a future policy to preserve EE demo compatibility despite this commit (it can result in highly patchy code if we keep doing it). But I love when I can preserve compatibility between versions… What is likely is that physics changes to the Doom code aren't going happen all the time, but only when we fix bugs (where indeed, adding demo_version sucks, because it means we keep crappy code) or do enhancements like here (where luckily it was only a need for these lines).

Well at least I have a possibility to test demos of previous a Eternity version against any of my regressions with slope physics.