Reverted the default player view pitch limit to ±32°
Sorry for this, but it caused more trouble than expected with how skies are rendered and especially how non-Doom games present themselves, so much that it caused one of the Eternity creators to think about reverting this himself.

However, I totally get why one would want more than 32° of pitch, so there'll definitely be some setting at some useful scope to increase it back to 45° or more.

We also have the group of PrBoom wads with tall skies extending differently in the first place, so the mess is bigger than it seems. It looks like Eviternity is relying on a dirty trick of using the MBF sky-transfer special everywhere (this is a hell of an overhead) and 'forgetting' to tag one sector, which takes advantage of an existing Eternity bug, which disables the unwanted panning of the sky upwards. This is REALLY messed up.