Confirm quit when closing window from the quit menu
If the user tries to close the game window while the quit menu is already
being displayed to the user, Eternity tries to open another quit menu on
top of it and ends up crashing. Instead, a more sensible option is to
interpret it as a confirmation, like e.g. Chocolate Doom or Woof! do.

This is done by turning the window close event into an `ev_quit` event
(also similarly to Chocolate Doom or Woof!). If the quit menu is not
open, this calls `MN_QuitDoom` from `MN_Responder`, and if it is, it is
interpreted as a confirmation inside `MN_PopupResponder`.

Link: https://github.com/chocolate-doom/chocolate-doom/commit/15f1bce94eec21fe9e11ee7e6745d01ccd4cfa70
Link: https://github.com/chocolate-doom/chocolate-doom/commit/5014caee2a000ffd682f04bcbec76265b2163335
Link: https://github.com/fabiangreffrath/woof/commit/bcf5410d3df9aa124e0418dc08f8f1d3137ddc6a