Attempt to move the slope sliding into P_TryMove
This prompted the definition of ceilingsector, in order to treat both ceiling and floor the same. However, it's bad to actually change only the displacement, because the velocity will remain high, causing player to suddenly spring to speed when away from slope, instead of continuing from the lower speed. Perhaps it's better to use P_XYMovement or whoever controls velocity.