Fix crappy bouncing of grenades on slopes. Fixes #654
I tried implementing geometrical bouncing off slopes for everything, as
if they're photons reflecting on plane surfaces. This is too unrealistic
for most cases and glitches too easily.

Instead now only fliers (like projectiles and not only) bounce like
this, while fallers bounce as if it's a normal floor (or stair), being
subject to the whole floor friction if they can't get enough momentum to
jump. Sorry for the change.