Added support for non-power-of-two flats
Flats which would get rejected in R_FindFlats because they didn't satisfy the power-of-two side (not necessarily square) are now classified as FLAT_NON_POWER_OF_2 (and can be flats), along with the template-function square ones and the generalized power-of-two category. Added new functions for this case, which have to use a modulus function that solves the stupid negative offset.